+++ frontend/app/src/components/dashboard/DashboardAssetTable.vue
@@ -1,19 +1,19 @@
 <template>
   <v-card>
-    <v-row no-gutters class="pa-3 secondary--text">
-      <v-toolbar-title class="font-weight-medium">
-        {{ title }}
-      </v-toolbar-title>
+    <v-card-title>
+      {{ title }}
       <v-spacer />
       <v-text-field
         v-model="search"
-        append-icon="mdi-magnify"
+        outlined
+        dense
+        prepend-inner-icon="mdi-magnify"
         :label="$t('dashboard_asset_table.search')"
-        class="pa-0 ma-0"
+        class="pa-0 ma-0 dashboard-asset-table__search"
         single-line
         hide-details
       />
-    </v-row>
+    </v-card-title>
     <v-card-text>
       <v-data-table
         class="dashboard-asset-table__balances"
@@ -26,11 +26,22 @@
         :footer-props="footerProps"
       >
         <template #header.usdValue>
-          {{
-            $t('dashboard_asset_table.headers.value', {
-              symbol: currencySymbol
-            })
-          }}
+          <div class="text-no-wrap">
+            {{
+              $t('dashboard_asset_table.headers.value', {
+                symbol: currencySymbol
+              })
+            }}
+          </div>
+        </template>
+        <template #header.price>
+          <div class="text-no-wrap">
+            {{
+              $t('dashboard_asset_table.headers.price', {
+                symbol: currencySymbol
+              })
+            }}
+          </div>
         </template>
         <template #item.asset="{ item }">
           <asset-details :asset="item.asset" />
@@ -46,6 +57,14 @@
             :value="item.usdValue"
           />
         </template>
+        <template #item.price="{ item }">
+          <amount-display
+            show-currency="symbol"
+            fiat-currency="USD"
+            tooltip
+            :value="prices[item.asset] ? prices[item.asset] : '-'"
+          />
+        </template>
         <template #item.percentage="{ item }">
           <percentage-display :value="percentage(item.usdValue)" />
         </template>
@@ -59,8 +78,11 @@
           </span>
         </template>
         <template v-if="balances.length > 0 && search.length < 1" #body.append>
-          <tr class="dashboard-asset-table__balances__total font-weight-medium">
-            <td colspan="2">{{ $t('dashboard_asset_table.total') }}</td>
+          <tr
+            v-if="$vuetify.breakpoint.smAndUp"
+            class="dashboard-asset-table__balances__total font-weight-medium"
+          >
+            <td colspan="3">{{ $t('dashboard_asset_table.total') }}</td>
             <td class="text-end">
               <amount-display
                 :fiat-currency="currencySymbol"
@@ -77,6 +99,29 @@
             </td>
             <td />
           </tr>
+          <tr v-else>
+            <td>
+              <v-row class="justify-space-between">
+                <v-col cols="auto" class="font-weight-medium">
+                  {{ $t('dashboard_asset_table.total') }}
+                </v-col>
+                <v-col cols="auto">
+                  <amount-display
+                    :fiat-currency="currencySymbol"
+                    :value="
+                      balances
+                        | aggregateTotal(
+                          currencySymbol,
+                          exchangeRate(currencySymbol),
+                          floatingPrecision
+                        )
+                    "
+                    show-currency="symbol"
+                  />
+                </v-col>
+              </v-row>
+            </td>
+          </tr>
         </template>
       </v-data-table>
     </v-card-text>
@@ -87,15 +132,17 @@
 import { default as BigNumber } from 'bignumber.js';
 import { Component, Prop, Vue } from 'vue-property-decorator';
 import { DataTableHeader } from 'vuetify';
-import { mapGetters } from 'vuex';
+import { mapGetters, mapState } from 'vuex';
 import { footerProps } from '@/config/datatable.common';
-import { AssetBalance } from '@/store/balances/types';
+import { CURRENCY_USD } from '@/data/currencies';
+import { AssetBalance, AssetPrices } from '@/store/balances/types';
 
 @Component({
   computed: {
     ...mapGetters('session', ['floatingPrecision', 'currencySymbol']),
     ...mapGetters('balances', ['exchangeRate']),
-    ...mapGetters('statistics', ['totalNetWorthUsd'])
+    ...mapGetters('statistics', ['totalNetWorthUsd']),
+    ...mapState('balances', ['prices'])
   }
 })
 export default class DashboardAssetTable extends Vue {
@@ -109,33 +156,48 @@ export default class DashboardAssetTable extends Vue {
   totalNetWorthUsd!: BigNumber;
   floatingPrecision!: number;
   currencySymbol!: string;
+  prices!: AssetPrices;
   exchangeRate!: (currency: string) => number | undefined;
 
   search: string = '';
 
   readonly footerProps = footerProps;
-  readonly headers: DataTableHeader[] = [
-    {
-      text: this.$tc('dashboard_asset_table.headers.asset'),
-      value: 'asset'
-    },
-    {
-      text: this.$tc('dashboard_asset_table.headers.amount'),
-      value: 'amount',
-      align: 'end'
-    },
-    {
-      text: this.$tc('dashboard_asset_table.headers.value'),
-      value: 'usdValue',
-      align: 'end'
-    },
-    {
-      text: this.$tc('dashboard_asset_table.headers.percentage'),
-      value: 'percentage',
-      align: 'end',
-      sortable: false
-    }
-  ];
+  get headers(): DataTableHeader[] {
+    return [
+      {
+        text: this.$tc('dashboard_asset_table.headers.asset'),
+        value: 'asset'
+      },
+      {
+        text: this.$t('dashboard_asset_table.headers.price', {
+          symbol: this.currencySymbol ?? CURRENCY_USD
+        }).toString(),
+        value: 'price',
+        align: 'end',
+        sortable: false
+      },
+      {
+        text: this.$tc('dashboard_asset_table.headers.amount'),
+        value: 'amount',
+        align: 'end',
+        width: '100%'
+      },
+      {
+        text: this.$t('dashboard_asset_table.headers.value', {
+          symbol: this.currencySymbol ?? CURRENCY_USD
+        }).toString(),
+        value: 'usdValue',
+        align: 'end'
+      },
+      {
+        text: this.$tc('dashboard_asset_table.headers.percentage'),
+        value: 'percentage',
+        align: 'end',
+        width: '120px',
+        sortable: false
+      }
+    ];
+  }
 
   percentage(value: BigNumber): string {
     return value.div(this.totalNetWorthUsd).multipliedBy(100).toFixed(2);
@@ -145,6 +207,10 @@ export default class DashboardAssetTable extends Vue {
 
 <style scoped lang="scss">
 .dashboard-asset-table {
+  &__search {
+    max-width: 450px;
+  }
+
   &__balances {
     &__total {
       &:hover {