+++ frontend/app/src/store/balances/getters.ts
@@ -24,6 +24,7 @@ import {
   GeneralAccount,
   KSM
 } from '@/typing/types';
+import { uniqueStrings } from '@/utils/array';
 import { assert } from '@/utils/assertions';
 import { bigNumberify, Zero } from '@/utils/bignumbers';
 import { assetSum } from '@/utils/calculation';
@@ -37,6 +38,7 @@ export interface BalanceGetters {
   exchanges: ExchangeInfo[];
   exchangeBalances: (exchange: string) => AssetBalance[];
   aggregatedBalances: AssetBalance[];
+  aggregatedAssets: string[];
   liabilities: AssetBalance[];
   manualBalanceByLocation: LocationBalance[];
   blockchainTotal: BigNumber;
@@ -435,5 +437,11 @@ export const getters: Getters<
       return match.type === 'ethereum token';
     }
     return false;
+  },
+  aggregatedAssets: (_, getters) => {
+    const liabilities = getters.liabilities.map(({ asset }) => asset);
+    const assets = getters.aggregatedBalances.map(({ asset }) => asset);
+    assets.push(...liabilities);
+    return assets.filter(uniqueStrings);
   }
 };